gtk_html_save needs an API with size_t instead of guint (#395553)
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@15395 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/gnome-utils/gnc-html.c
@@ -1291,7 +1291,7 @@ gnc_html_copy(gnc_html *html)
 static gboolean 
 raw_html_receiver (gpointer     engine,
                    const gchar *data,
-                   guint        len,
+                   size_t        len,
                    gpointer     user_data)
 {
   FILE *fh = (FILE *) user_data;
